refactor: 重构 stores
