+++ src/locale/index.ts
@@ -1,5 +1,6 @@
-import { addMessages, init, locale } from 'svelte-i18n';
+import { addMessages, init } from 'svelte-i18n';
 import { get } from 'svelte/store';
+import { lang } from '../stores';
 import { SupportedLang } from '../utils/enum';
 
 import zhCN from './lang/zh-CN'
@@ -10,18 +11,6 @@ export function setupI18n() {
     addMessages(SupportedLang.zhTW, zhTW);
     init({
         fallbackLocale: SupportedLang.zhCN,
-        initialLocale: localStorage.getItem('lang'),
+        initialLocale: get(lang),
     });
 }
-
-export const lang = {
-    subscribe: (run: (value: SupportedLang) => void) => { return locale.subscribe(run) },
-    switchTo: (lang: SupportedLang) => { locale.set(lang) }
-}
-
-export const changeLang = (target: SupportedLang) => {
-    if (target != get(lang)) {
-        lang.switchTo(target);
-        localStorage.setItem('lang', target);
-    }
-};
