+++ src/stores.ts
@@ -2,19 +2,24 @@
  * Stores和全局变量
  * @author wniko
  */
-import { writable } from 'svelte/store';
+import { Writable, writable } from 'svelte/store';
 import type { MapPoint } from './utils/typings';
 import { ConvertType } from 'zhconvertor';
 import { persist } from './utils/persist';
+import { locale } from 'svelte-i18n';
+import type { SupportedLang } from './utils/enum';
 
 // Stores
 
 /** 是否是管理员Mode的store */
 export let isAdminModeStore = writable(false);
 
-/** 当前内容语言 */
-// 空：不转换
+/** 界面语言 */
+export let lang = persist<SupportedLang>(locale as Writable<SupportedLang>, 'lang');
+
+/** 内容语言转换 */
 export let convertTargetStore = persist(writable<ConvertType>(ConvertType.dont), 'convertTarget');
+
 
 //全局变量
